backend/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
aniso8601==9.0.1
argon2-cffi==21.1.0
cffi==1.15.0
click==8.0.3; python_version >= '3.6'
flask-cors==3.0.10
flask-jwt-extended==4.3.1
flask-marshmallow==0.14.0
flask-restful==0.3.9
flask-sqlalchemy==2.5.1
flask==2.0.2
greenlet==1.1.2
itsdangerous==2.0.1; python_version >= '3.6'
jinja2==3.0.2; python_version >= '3.6'
markupsafe==2.0.1; python_version >= '3.6'
marshmallow-sqlalchemy==0.26.1
marshmallow==3.13.0
psycopg2-binary==2.9.1
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyjwt==2.3.0
pytz==2021.3
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==1.4.25
uwsgi==2.0.20
werkzeug==2.0.2; python_version >= '3.6'
